Remove webpack
