+++ package.json
@@ -1,10 +1,8 @@
 {
   "dependencies": {
     "@hotwired/stimulus": "^3.0.1",
-    "@hotwired/stimulus-webpack-helpers": "^1.0.1",
     "@hotwired/turbo-rails": "^7.1.3",
     "@rails/actioncable": "^7.0.2-2",
-    "@rails/webpacker": "^5.4.3",
     "bootstrap-notify": "^3.1.3",
     "chart.js": "^3.8.0",
     "chartkick": "^4.2.0",
@@ -15,8 +13,7 @@
     "vue-turbolinks": "^2.2.2"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "webpack-dev-server": "^4.5.0"
+    "@babel/plugin-proposal-class-properties": "^7.7.4"
   },
   "scripts": {
     "build": "esbuild app/javascript/*.* --bundle --sourcemap --outdir=app/assets/builds --public-path=assets"
